requirements.txt
openai
Pillow
numpy 